### 2019-04-07 #1367 @alexhanna
@Yury-Shevchenko can you suggest some possible reviewers for this piece? You can see the list of possible reviewers earlier in this thread.

Can you also ensure that your paper.md file is formatted correctly? Our build script seems to be failing. The paper.md format details are in the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html).
