+++ paper.md
@@ -12,7 +12,7 @@ affiliations:
   -	name: University of Mannheim, Germany
 index: 1
 date: 26 March 2019
-bibliography: bibliography.bib
+bibliography: paper.bib
 ---
 
 # Summary
