+++ paper.md
@@ -25,13 +25,13 @@ Figure 1: The movie task
 
 The multi-attribute decision paradigm in a tabular format was described as "Mouselab" in the book by Payne, Bettman and Johnson "The adaptive decision maker" [@Payne1993]. Since then, it has been a popular approach for investigating decision processes. Following the general guidelines, researchers developed their own laboratory versions of Mouselab tailored to their needs [@Jasper2002, @Reisen2008]. Nevertheless, due to differences in software and programming languages, it was difficult to share tasks within the research community.
 
-Willemsen and Johnson developed MouselabWEB as an online version of Mouselab [@Willemsen2010a]. MouselabWEB supports HTML and JavaScript and has been available online for many years since its release [@DanGoldstein2006]. However, there is a clear need for modern applications that help researchers design and publish their own Mouselab tasks online with greater flexibility and usability.
+Willemsen and Johnson developed MouselabWEB as an online version of Mouselab [@Willemsen2010]. MouselabWEB supports HTML and JavaScript and has been available online for many years since its release [@DanGoldstein2006]. However, there is a clear need for modern applications that help researchers design and publish their own Mouselab tasks online with greater flexibility and usability.
 
 We present the Multi-Attribute Decision (MAD) builder for creating a Mouselab task in an online experiment. The MAD builder is available as a template in [lab.js](https://labjs.felixhenninger.com/), free and open-source software. The task builder is also accessible at [Open Lab](https://open-lab.online), an online platform for deployment of experiments.
 
 MAD builder gives a researcher control over many parameters such as the display of the task, randomization algorithms, feedback options, and presentation of cues as text or images. Saved as a JSON file, MAD builder file can be easily archived or shared among researchers. The user manual for the task is available on the [GitHub page](https://github.com/Yury-Shevchenko/mad). The manual describes how to customize the task by changing task parameters. It presents four usage scenarios with different types of content and information presentation strategies. At the end, the manual shows how to work with the data generated in an experiment.
 
-The tabular format of the multi-attribute decision task has prevailed in previous process-tracing studies [@Willemsen2010a]. The MAD builder allows the construction of tables of different sizes and positioning options either in the top row or in the left column. The builder has been already used for the design of experimental tasks in several studies on decision-making [@ScharfS, @Shevchenko2018].
+The tabular format of the multi-attribute decision task has prevailed in previous process-tracing studies [@Willemsen2010a]. The MAD builder allows the construction of tables of different sizes and positioning options either in the top row or in the left column. The builder has been already used for the design of experimental tasks in several studies on decision-making [@Scharf2019, @Shevchenko2018].
 
 The presentation of information in tables, however, limits the scope of research questions to the analysis of how people deal with structured information. Researchers may be interested in learning how people behave in other contexts that employ a more ecologically valid design in which information about cues is scattered around and embedded in the environment. New research questions can be asked in this case, such as how do people organize available information [@Ettlin2015]. The MAD builder allows the user to fully customize the display of options and cues by positioning them in any place on the screen. In this way, participants can interact with images of objects and examine their attributes and values by clicking on them (see Figure 2).
 
