+++ paper.bib
@@ -23,7 +23,7 @@
   year = {2008}
 }
 
-@incollection{Willemsen2010a,
+@incollection{Willemsen2010,
   abstract = {A major task for all models of cognition is the development of representations and processes that both predict outcomes and are consistent with what we know about underlying cognitive structure. It is not surprising, therefore, that a single set of outputs can result from different ...},
   author = {Willemsen, Martijn C. and Johnson, Eric J.},
   booktitle = {A handbook of process tracing methods for decision research},
@@ -33,7 +33,7 @@
   year = {2011}
 }
 
-@article{ScharfS.E.WiegelmannM.&Broder2019,
+@article{Scharf2019,
   author = {Scharf, Sophie and Wiegelmann, Monika and Br{\"{o}}der, Arndt},
   journal = {Manuscript submitted for publication.},
   title = {{Information Search in Everyday Decisions: The Generalizability of the Attraction Search Effect.}},
@@ -48,16 +48,6 @@
   journal = {Behavior Research Methods, Instruments, and Computers},
   title = {{MouseTrace: A better mousetrap for catching decision processes}},
   year = {2002}
-}
-
-@incollection{Willemsen2010a,
-  abstract = {A major task for all models of cognition is the development of representations and processes that both predict outcomes and are consistent with what we know about underlying cognitive structure. It is not surprising, therefore, that a single set of outputs can result from different ...},
-  author = {Willemsen, Martijn C. and Johnson, Eric J.},
-  booktitle = {A handbook of process tracing methods for decision research},
-  editor = {Schulte-Mecklenbeck, M. and K{\"{u}}hberger, A. and Ranyard, R.},
-  isbn = {978-1848728646},
-  title = {{Visiting the Decision Factory: Observing Cognition with MouselabWEB and other Information Acquisition Methods.}},
-  year = {2011}
 }
 
 @article{Shevchenko2018,
