### 2019-04-07 #1367 @Yury-Shevchenko
Dear @alexhanna , I've fixed formatting in the paper.md and paper.bib files, hope it will work now.
As a reviewer I would suggest @andytwoods based on the language (JavaScript) and the domain (psychology).
