### 2019-04-07 #1367 @kyleniemeyer
@Yury-Shevchenko I think the issue may be in the header of your `paper.md`. Could you modify the author and affiliation part to look like:
```
authors:
  - name: Yury Shevchenko
    orcid: 0000-0002-3397-8465
    affiliation: 1
affiliations:
  - name: University of Mannheim, Germany
    index: 1
```
