+++ paper.md
@@ -5,11 +5,12 @@ tags:
   -	online experiment
   -	JavaScript
 authors:
-  -	name: Yury Shevchenko
-orcid: 0000-0002-3397-8465
-affiliation: 1
+  - name: Yury Shevchenko
+    orcid: 0000-0002-3397-8465
+    affiliation: 1
 affiliations:
-  -	name: University of Mannheim, Germany
+  - name: University of Mannheim, Germany
+    index: 1
 index: 1
 date: 26 March 2019
 bibliography: paper.bib
@@ -31,7 +32,7 @@ We present the Multi-Attribute Decision (MAD) builder for creating a Mouselab ta
 
 MAD builder gives a researcher control over many parameters such as the display of the task, randomization algorithms, feedback options, and presentation of cues as text or images. Saved as a JSON file, MAD builder file can be easily archived or shared among researchers. The user manual for the task is available on the [GitHub page](https://github.com/Yury-Shevchenko/mad). The manual describes how to customize the task by changing task parameters. It presents four usage scenarios with different types of content and information presentation strategies. At the end, the manual shows how to work with the data generated in an experiment.
 
-The tabular format of the multi-attribute decision task has prevailed in previous process-tracing studies [@Willemsen2010a]. The MAD builder allows the construction of tables of different sizes and positioning options either in the top row or in the left column. The builder has been already used for the design of experimental tasks in several studies on decision-making [@Scharf2019, @Shevchenko2018].
+The tabular format of the multi-attribute decision task has prevailed in previous process-tracing studies [@Willemsen2010]. The MAD builder allows the construction of tables of different sizes and positioning options either in the top row or in the left column. The builder has been already used for the design of experimental tasks in several studies on decision-making [@Scharf2019, @Shevchenko2018].
 
 The presentation of information in tables, however, limits the scope of research questions to the analysis of how people deal with structured information. Researchers may be interested in learning how people behave in other contexts that employ a more ecologically valid design in which information about cues is scattered around and embedded in the environment. New research questions can be asked in this case, such as how do people organize available information [@Ettlin2015]. The MAD builder allows the user to fully customize the display of options and cues by positioning them in any place on the screen. In this way, participants can interact with images of objects and examine their attributes and values by clicking on them (see Figure 2).
 
