### 2019-04-07 3adc8ca
Removed extra "index" entry from paper
