### 2019-04-07 #1367 @kyleniemeyer
@Yury-Shevchenko I just submitted a second PR, which makes GitHub preview the paper.md file the way it does for the rest of our papers. Let's give that a shot—not exactly sure what the issue was.
