### 2019-04-07 #1367 @kyleniemeyer
@Yury-Shevchenko @alexhanna success! It looks like there still might be some content things to fix regarding the figures, but at least it's compiling now.
