### 2019-04-07 #1367 @Yury-Shevchenko
If I may, I'll fix the figures, and recompile the pdf.
