### 2019-04-24 #1409 @whedon
[REVIEW]: Multi-Attribute Decision task builder for online experiments
**Submitting author:** @Yury-Shevchenko (<a href="http://orcid.org/0000-0002-3397-8465">Yury Shevchenko</a>)
**Repository:** <a href="https://github.com/Yury-Shevchenko/mad" target ="_blank">https://github.com/Yury-Shevchenko/mad</a>
**Version:** 1.0.0
