### 2019-05-23 Yury-Shevchenko/mad#3 @u01ai11
Lab.js example links in README.md do not work
When I click through on the Lab.js links in the example they do not work! I just get taken to the Lab.js builder site with an empty project.

I can see the lab.js url contains open-lab project, but nothing is loaded from there.
