### 2019-05-23 Yury-Shevchenko/mad#3 @Yury-Shevchenko
Hello Alex! First, thanks a lot for reviewing the paper.
I've just checked - the links work with me.
Do you see a pop-up dialogue such as
"Would you like to load the study from https://open-lab.online/labjs/5ca3648b8ef62535954f9c50?"
It might be that dialogues are blocked in your browser. I would try different browser (specially, if you on a mac).
