### 2019-05-23 Yury-Shevchenko/mad#3 @u01ai11
@Yury-Shevchenko Thanks -- it does work if I remove all my plugins from Chrome.

I'll close the issue - but it might be good to put a note in your README so users will remember to do turn-off blockers etc.
