### 2019-05-24 #1409 @u01ai11
This issue -- problems with opening the examples in Lab.js was resolved: https://github.com/Yury-Shevchenko/mad/issues/3
