### 2019-05-24 Yury-Shevchenko/mad#4 @u01ai11
I can't work out how to edit parameters in Open Lab
In the README it says:
"The second option is to open the task in Open Lab, an online platform for deployment of experiments. To modify the task in lab.js or in Open Lab, you have to edit its parameters"

When I try and do this in Open Lab via the link (and after signing up -- this should also be mentioned in your documentation), I can only preview the example, I don't seem to have access to the JSON parameters -- as I can do easily in in Lab.js.
