### 2019-05-24 #1409 @u01ai11
In going through the examples, I can't find a way to edit the parameters in Open Lab, just Lab.js. The documentation should also mention that you need an account to view the example templates. https://github.com/Yury-Shevchenko/mad/issues/4
