### 2019-05-24 Yury-Shevchenko/mad#5 @u01ai11
There is no release version
According to JOSS review checklist I need to check that your release version matches the repository version. But you have not made a release on GitHub. Please do this so I can check that box!

https://help.github.com/en/articles/creating-releases
