### 2019-05-24 #1409 @u01ai11
The author has not made a release for the github repo

https://github.com/Yury-Shevchenko/mad/issues/5
