### 2019-05-24 Yury-Shevchenko/mad#4 @Yury-Shevchenko
Dear Alex,
Editing the parameters on Open Lab is realized through the study projects.  The user can change parameters of the task for each project. In this way, the original task remains unchanged, but users can create projects in which they use different versions of the task.
The documentation can be found here:
https://open-lab.online/docs/project
I will add the information about these steps to the README file.
