### 2019-05-24 Yury-Shevchenko/mad#4 @u01ai11
Thanks!

It would be good to go through the steps briefly (i.e. Signup -> New Study -> add MAD task in the study -> edit paramaters).
