### 2019-05-24 Yury-Shevchenko/mad#6 @u01ai11
Community Guidelines for contribution
As per JOSS guidelines, there needs to be community guidelines:

" third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

This could be added to the wiki or the README
