### 2019-05-24 #1409 @u01ai11
Cannot check the community guidelines box

https://github.com/Yury-Shevchenko/mad/issues/6
