### 2019-05-24 Yury-Shevchenko/mad#7 @u01ai11
Missing doi from one reference
In paper.md there is a missing doi for the entry:

"Reisen, N., Hoffrage, U., & Mast, F. W. (2008). Identifying decision strategies in a consumer choice situation. Judgment and decision making, 3(8), 641–658."
