### 2019-05-24 #1409 @u01ai11
Missing a doi for one paper in the paper.md file

https://github.com/Yury-Shevchenko/mad/issues/7
