### 2019-05-24 #1409 @u01ai11
I can't tick the checkbox, as there are no automated tests or steps outlined to verify the functioning of the software is as expected.
Given that this is a template for working on Lab.js and Open Lab, I am not sure what this would look like, or if this is necessary?
