### 2019-05-24 #1409 @u01ai11
OK - I have completed the review for this.

I think this is a great tool for creating 'Mouselab' tasks quickly using an existing web-based builder (Lab.js) and a SaaS platform (Open Lab).

The documentation (in the README file) is clear, and the examples are helpful for understanding the functionality of the templates. One suggestion (rather than acceptance blocker) is to create a broken down by heading documentation using the GitHub's repo's wiki -- or some internal links on the README so you can skip to the relevant part of the doc easily.

Generally speaking, probably beyond the scope of acceptance criteria of JOSS, I am not sure where the line between complete 'Software' and a template experiment built in other software lies -- and where this particular submission falls?

But I do think this tool would benefit the research community who work using this particular type of task -- and as such, it should be published.

In summary, the things that need to be resolved for the checklist to be complete are:
- Reference DOI for one citation needs to be added.
- A release on Github needs to be made.
- Community guidelines for contributing etc need to be made.
- Some steps for verifying correct functioning need to be detailed, or @alexhanna agree that testing is not applicable in this instance.
