### 2019-05-27 Yury-Shevchenko/mad#7 @Yury-Shevchenko
This article does not have a doi. There is no doi on CrossRef or in other publications that cite this article. The article has been published in Judgment and Decision Making Journal (http://journal.sjdm.org/). The URL of the article is http://journal.sjdm.org/8801/jdm8801.html
