+++ paper.bib
@@ -20,6 +20,7 @@
   pages = {641--658},
   title = {{Identifying decision strategies in a consumer choice situation}},
   volume = {3},
+  url = {http://journal.sjdm.org/8801/jdm8801.html},
   year = {2008}
 }
 
