### 2019-05-27 Yury-Shevchenko/mad#6 @Yury-Shevchenko
Community guidelines were added to the README file.
