### 2019-05-27 Yury-Shevchenko/mad#4 @Yury-Shevchenko
The information about opening a task in Open lab and lab.js was added to the Readme file.
