### 2019-05-27 #1409 @Yury-Shevchenko
Dear Alex,
Thank you for your review and recognition of the merits of the multi-attribute decision builder.
To address your comments, I've added more information on how to open the task in Open Lab and lab.js in the documentation. Community guidelines have also been added.
As for your recommendation on wiki pages, this is definitely a good idea that I will follow in the future.
Concerning your comment on the boundary between software and a template, I understand your point of view, but I would also like to stress that the multi-attribute decision builder has its own independent functionality. It is about 700 lines of JavaScript code written to support the creation of a task with parameter values. So my point is that the builder is not just a template that was created in another program. Maybe the choice of the word "template" was a bit misleading, but the motivation behind it was to show a user that he/she can easily edit and customize the task.
I have updated the version on GitHub as well, and look forward to your feedback.
Best,
Yury
