### 2019-05-28 #1409 @u01ai11
Dear Yury,

Thanks for making the changes I suggested. I have now been able to tick off all the boxes (other than the testing box) and as far as I am concerned, I would be pleased for it to be accepted in the current form.

Also, thanks for addressing my comment, it was more a point of thought rather than a criticism of the submission -- as I said before, I think this software will be very helpful for those building MAD tasks (especially those who are new to online research, or the type of task itself).

Best wishes,
Alex
