### 2019-05-28 #1409 @u01ai11
> whedon set v1.0 as version

I think this needs to be done, as the repo release was created!
