### 2019-06-16 #1409 @alexhanna
Great.  @Yury-Shevchenko, the last thing we need for this submission to make an software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the Zenodo/figshare archive, please make sure that:
