### 2019-06-17 #1409 @alexhanna
@Yury-Shevchenko looks like one of the references is missing a DOI. Can you check on this?
