### 2019-06-17 #1409 @Yury-Shevchenko
Dear @alexhanna I have not found the doi for this article. There is no doi on CrossRef or in other publications that cite this article. The article has been published in Judgment and Decision Making Journal (http://journal.sjdm.org/). The URL of the article is http://journal.sjdm.org/8801/jdm8801.html. I could not open the reference (https://doi.org/10.1037/e722292011-085) suggested by the search above.
