### 2019-06-17 #1409 @alexhanna
@openjournals/joss-eics - what should happen in this case?
