### 2019-06-17 #1409 @kyleniemeyer
@alexhanna whedon's DOI checker isn't perfect, so if manual inspection finds that there is no issue, then we can ignore it.
