### 2019-06-18 #1409 @danielskatz
👋 @Yury-Shevchenko - I see a few small wording issues in the paper, which I have addressed in https://github.com/Yury-Shevchenko/mad/pull/8
