### 2019-06-18 #1409 @Yury-Shevchenko
Dear @danielskatz Thanks a lot for that! I have merged the changes.
